Added rest of item changes from v1.3

+++ src/main/java/net/soulsweaponry/items/bow/KrakenSlayer.java
@@ -47,7 +47,7 @@ public class KrakenSlayer extends ModdedBow {
 
     @Override
     public int getPullTime() {
-        return ConfigConstructor.kraken_slayer_pull_time_ticks;
+        return (int) ConfigConstructor.kraken_slayer_pull_time_ticks;
     }
 
     @Override
