+++ src/main/java/net/soulsweaponry/items/scythe/ForlornScythe.java
@@ -31,7 +31,7 @@ public class ForlornScythe extends SoulHarvestingItem implements GeoItem {
     private final AnimatableInstanceCache factory = GeckoLibUtil.createInstanceCache(this);
 
     public ForlornScythe(ToolMaterial toolMaterial, Settings settings) {
-        super(toolMaterial, ConfigConstructor.forlorn_scythe_damage, ConfigConstructor.forlorn_scythe_attack_speed, settings);
+        super(toolMaterial, (int) ConfigConstructor.forlorn_scythe_damage, ConfigConstructor.forlorn_scythe_attack_speed, settings);
         this.addTooltipAbility(TooltipAbilities.SOUL_RELEASE_WITHER);
     }
