+++ src/main/java/net/soulsweaponry/items/scythe/SoulReaper.java
@@ -39,7 +39,7 @@ public class SoulReaper extends SoulHarvestingItem implements GeoItem, ISummonAl
     private final AnimatableInstanceCache factory = GeckoLibUtil.createInstanceCache(this);
 
     public SoulReaper(ToolMaterial toolMaterial, Settings settings) {
-        super(toolMaterial, ConfigConstructor.soul_reaper_damage, ConfigConstructor.soul_reaper_attack_speed, settings);
+        super(toolMaterial, (int) ConfigConstructor.soul_reaper_damage, ConfigConstructor.soul_reaper_attack_speed, settings);
         this.addTooltipAbility(TooltipAbilities.SOUL_RELEASE);
     }
 
@@ -136,7 +136,7 @@ public class SoulReaper extends SoulHarvestingItem implements GeoItem, ISummonAl
 
     @Override
     public int getMaxSummons() {
-        return ConfigConstructor.soul_reaper_summoned_allies_cap;
+        return (int) ConfigConstructor.soul_reaper_summoned_allies_cap;
     }
 
     @Override
